feat : queryDSL config 추가, 파일 구조 개선

+++ backend/src/main/java/com/ssafy/ptpt/db/entity/Evaluation.java
@@ -20,7 +20,7 @@ import org.springframework.boot.autoconfigure.domain.EntityScan;
 public class Evaluation {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int evaluationNo;
+    private Long evaluationNo;
 
     private String presentationName;
     private int delivery;
